feat: Create & Impl GalleryNumberRepository

+++ app/src/main/java/com/vtsb/hipago/data/datasource/remote/entity/GalleryNumber.kt
@@ -2,5 +2,5 @@ package com.vtsb.hipago.data.datasource.remote.entity
 
 data class GalleryNumber(
     val numberList: List<Int>,
-    val length: Long
+    val length: Int
 )
